requirements.txt
streamlit==1.54.0
fastapi==0.128.6
pydantic==2.12.5
uvicorn==0.40.0
mysql-connector-python==9.6.0
pandas==2.3.3
requests==2.32.5
pytest==9.0.2